chore: consolidate runner codes in bench_runner directory
